ggincerta: Extend 'ggplot2' with Layers and Scales for Spatial Uncertainty Visualization

Provide specialized 'ggplot2' layers and scales for spatial uncertainty visualization, including bivariate choropleth maps, pixel maps, glyph maps, and exceedance probability maps.

See also

Author

Maintainer: Xueqi Ma maggiexma07@gmail.com [copyright holder]

Authors:

Xueqi Ma maggiexma07@gmail.com [copyright holder]

Emi Tanaka dr.emi.tanaka@gmail.com (ORCID) [thesis advisor]

Other contributors:

Weihao Li llreczx@gmail.com (ORCID) [thesis advisor]

Quan Vu [thesis advisor]

Francis Hui fhui28@gmail.com (ORCID) [thesis advisor]